include dependencies reworked
All header files in the main OpenGD project now only include the files they actually need and forward declare the rest. This also then includes the needed files in the .cpp files wherever needed. This should help improve compile times as well as reduce how often specific files need to be recompiled when changes get made as the dependencies are as minimal as possible.